+++ airbyte-integrations/bases/base-python-test/base_python_test/__init__.py
@@ -22,8 +22,9 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-from .test_iface import StandardSourceTestIface
+from .test_iface import StandardSourceTestIface, DefaultStandardSourceTest
 
 __all__ = [
     "StandardSourceTestIface",
+    "DefaultStandardSourceTest",
 ]
