+++ airbyte-integrations/bases/base-python-test/base_python_test/test_iface.py
@@ -24,21 +24,24 @@ SOFTWARE.
 
 from typing import List
 
-from airbyte_protocol import AirbyteCatalog, ConnectorSpecification
+import json
+import pkgutil
+
+from airbyte_protocol import AirbyteCatalog, ConfiguredAirbyteCatalog, ConnectorSpecification
 
 
-class StandardSourceTestIface(object):
+class StandardSourceTestIface:
     def __init__(self):
         pass
 
     def get_spec(self) -> ConnectorSpecification:
-        raise Exception("Not Implemented")
+        raise NotImplementedError
 
     def get_config(self) -> object:
-        raise Exception("Not Implemented")
+        raise NotImplementedError
 
     def get_catalog(self) -> AirbyteCatalog:
-        raise Exception("Not Implemented")
+        raise NotImplementedError
 
     def get_regex_tests(self) -> List[str]:
         return []
@@ -48,3 +51,20 @@ class StandardSourceTestIface(object):
 
     def tear_down(self) -> None:
         pass
+
+
+class DefaultStandardSourceTest(StandardSourceTestIface):
+    SPEC_FILENAME = "spec.json"
+    CONFIG_FILENAME = "config.json"
+    CONFIGURED_CATALOG_FILENAME = "configured_catalog.json"
+
+    def get_spec(self) -> ConnectorSpecification:
+        raw_spec = pkgutil.get_data(self.__class__.__module__.split(".")[0], self.SPEC_FILENAME)
+        return ConnectorSpecification.parse_obj(json.loads(raw_spec))
+
+    def get_config(self) -> object:
+        return json.loads(pkgutil.get_data(self.__class__.__module__.split(".")[0], self.CONFIG_FILENAME))
+
+    def get_catalog(self) -> ConfiguredAirbyteCatalog:
+        raw_catalog = pkgutil.get_data(self.__class__.__module__.split(".")[0], self.CONFIGURED_CATALOG_FILENAME)
+        return ConfiguredAirbyteCatalog.parse_obj(json.loads(raw_catalog))
