+++ airbyte-integrations/bases/base-singer/base_singer/source.py
@@ -22,9 +22,9 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-from typing import Generator
+from typing import Generator, Type
 
-from airbyte_protocol import AirbyteCatalog, AirbyteMessage, ConfiguredAirbyteCatalog
+from airbyte_protocol import AirbyteCatalog, AirbyteMessage, ConfiguredAirbyteCatalog, AirbyteConnectionStatus, Status
 from base_python import AirbyteLogger, ConfigContainer, Source
 
 from .singer_helpers import SingerHelper
@@ -36,7 +36,7 @@ class SingerSource(Source):
         Returns the command used to run discovery in the singer tap. For example, if the bash command used to invoke the singer tap is `tap-postgres`,
         and the config JSON lived in "/path/config.json", this method would return "tap-postgres --config /path/config.json"
         """
-        raise Exception("Not Implemented")
+        raise NotImplementedError
 
     def read_cmd(self, logger: AirbyteLogger, config_path: str, catalog_path: str, state_path: str = None) -> str:
         """
@@ -44,7 +44,7 @@ class SingerSource(Source):
         and the config JSON lived in "/path/config.json", and the catalog was in "/path/catalog.json",
         this method would return "tap-postgres --config /path/config.json --catalog /path/catalog.json"
         """
-        raise Exception("Not Implemented")
+        raise NotImplementedError
 
     def discover(self, logger: AirbyteLogger, config_container: ConfigContainer) -> AirbyteCatalog:
         """
@@ -68,3 +68,52 @@ class SingerSource(Source):
 
         read_cmd = self.read_cmd(logger, config_container.rendered_config_path, selected_singer_catalog_path, state_path)
         return SingerHelper.read(logger, read_cmd)
+
+
+class BaseSingerSource(SingerSource):
+    def discover_cmd(self, logger: AirbyteLogger, config_path: str) -> str:
+        return f"{self.tap_cmd} --config {config_path} --discover"
+
+    def read_cmd(self, logger: AirbyteLogger, config_path: str, catalog_path: str, state_path: str = None) -> str:
+        args = {"--config": config_path, "--catalog": catalog_path, "--state": state_path}
+        cmd = " ".join([f"{k} {v}" for k, v in args.items() if v is not None])
+
+        return f"{self.tap_cmd} {cmd}"
+
+    def check(self, logger: AirbyteLogger, config_container: ConfigContainer) -> AirbyteConnectionStatus:
+        try:
+            json_config = config_container.rendered_config
+            self.try_connect(logger, json_config)
+        except self.api_error as err:
+            logger.error("Exception while connecting to the %s: %s", self.tap_name, str(err))
+            # this should be in UI
+            error_msg = (
+                f"Unable to connect to the {self.tap_name} with the provided credentials. "
+                f"Error: {err}"
+            )
+            return AirbyteConnectionStatus(status=Status.FAILED, message=error_msg)
+
+        return AirbyteConnectionStatus(status=Status.SUCCEEDED)
+
+    def try_connect(self, logger: AirbyteLogger, config: dict):
+        """ Test provided credentials, raises self.api_error if something goes wrong
+        """
+        raise NotImplementedError
+
+    @property
+    def api_error(self) -> Type[Exception]:
+        """ Class/Base class of the exception that will be thrown if the tap is misconfigured or service unavailable
+        """
+        raise NotImplementedError
+
+    @property
+    def tap_cmd(self) -> str:
+        """ Tap command
+        """
+        raise NotImplementedError
+
+    @property
+    def tap_name(self) -> str:
+        """ Tap name
+        """
+        raise NotImplementedError
