+++ airbyte-integrations/connectors/source-intercom-singer/source_intercom_singer/source.py
@@ -21,44 +21,11 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
+from typing import Mapping, Any
 
-from airbyte_protocol import AirbyteConnectionStatus, Status
 from base_python import AirbyteLogger, ConfigContainer
-from base_singer import SingerSource
+from base_singer import BaseSingerSource
 from tap_intercom.client import IntercomClient, IntercomError
-
-
-class BaseSingerSource(SingerSource):
-    tap_cmd = None
-    tap_name = None
-    api_error = Exception
-
-    def discover_cmd(self, logger: AirbyteLogger, config_path: str) -> str:
-        return f"{self.tap_cmd} --config {config_path} --discover"
-
-    def read_cmd(self, logger: AirbyteLogger, config_path: str, catalog_path: str, state_path: str = None) -> str:
-        args = {"--config": config_path, "--catalog": catalog_path, "--state": state_path}
-        cmd = " ".join([f"{k} {v}" for k, v in args.items() if v is not None])
-
-        return f"{self.tap_cmd} {cmd}"
-
-    def check(self, logger: AirbyteLogger, config_container: ConfigContainer) -> AirbyteConnectionStatus:
-        try:
-            json_config = config_container.rendered_config
-            self.try_connect(logger, json_config)
-        except IntercomError as err:
-            logger.error("Exception while connecting to the %s: %s", self.tap_name, str(err))
-            # this should be in UI
-            error_msg = (
-                f"Unable to connect to the {self.tap_name} with the provided credentials. "
-                f"Error: {err}"
-            )
-            return AirbyteConnectionStatus(status=Status.FAILED, message=error_msg)
-
-        return AirbyteConnectionStatus(status=Status.SUCCEEDED)
-
-    def try_connect(self, logger: AirbyteLogger, config: dict):
-        raise NotImplementedError
 
 
 class SourceIntercomSinger(BaseSingerSource):
@@ -66,7 +33,7 @@ class SourceIntercomSinger(BaseSingerSource):
     tap_name = "Intercom API"
     api_error = IntercomError
 
-    def transform_config(self, raw_config):
+    def transform_config(self, raw_config) -> Mapping[str, Any]:
         return {
             "user_agent": "airbyte",
             "access_token": raw_config["access_token"],
